read file in 1kb chunks for SHA
